[REF] messages: update ticket system and apply timeout to actions
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@62161 b456876b-0849-0410-b77d-98878d47e9d5

+++ messu-sent.php
@@ -14,45 +14,47 @@ $messulib = TikiLib::lib('message');
 $access->check_user($user);
 $access->check_feature('feature_messages');
 $access->check_permission('tiki_p_messages');
+$access->checkAuthenticity();
 $maxRecords = $messulib->get_user_preference($user, 'maxRecords', 20);
+
+if ($access->ticketMatch()) {
 // Delete messages if the delete button was pressed
-if (isset($_REQUEST["delete"]) && isset($_REQUEST["msg"])) {
-	check_ticket('messu-sent');
-	foreach (array_keys($_REQUEST["msg"]) as $msg) {
-		$messulib->delete_message($user, $msg, 'sent');
-	}
-}
-// Archive messages if the archive button was pressed
-if (isset($_REQUEST["archive"]) && isset($_REQUEST["msg"])) {
-	check_ticket('messu-sent');
-	$tmp = $messulib->count_messages($user, 'archive');
-	foreach (array_keys($_REQUEST["msg"]) as $msg) {
-		if (($prefs['messu_archive_size'] > 0) && ($tmp >= $prefs['messu_archive_size'])) {
-			$smarty->assign('msg', tra("Archive is full. Delete some messages from archive first."));
-			$smarty->display("error.tpl");
-			die;
-		}
-		$messulib->archive_message($user, $msg, 'sent');
-		$tmp++;
-	}
-}
-// Download messages if the download button was pressed
-if (isset($_REQUEST["download"])) {
-	check_ticket('messu-sent');
-	// if message ids are handed over, use them:
-	if (isset($_REQUEST["msg"])) {
+	if (isset($_REQUEST["delete"]) && isset($_REQUEST["msg"])) {
 		foreach (array_keys($_REQUEST["msg"]) as $msg) {
-			$tmp = $messulib->get_message($user, $msg, 'sent');
-			$items[] = $tmp;
+			$messulib->delete_message($user, $msg, 'sent');
 		}
-	} else {
-		$items = $messulib->get_messages($user, 'sent', '', '', '');
 	}
-	$smarty->assign_by_ref('items', $items);
-	header("Content-Disposition: attachment; filename=tiki-msg-sent-" . time("U") . ".txt ");
-	$smarty->display('messu-download.tpl', null, null, null, 'application/download');
-	die;
+// Archive messages if the archive button was pressed
+	if (isset($_REQUEST["archive"]) && isset($_REQUEST["msg"])) {
+		$tmp = $messulib->count_messages($user, 'archive');
+		foreach (array_keys($_REQUEST["msg"]) as $msg) {
+			if (($prefs['messu_archive_size'] > 0) && ($tmp >= $prefs['messu_archive_size'])) {
+				$smarty->assign('msg', tra("Archive is full. Delete some messages from archive first."));
+				$smarty->display("error.tpl");
+				die;
+			}
+			$messulib->archive_message($user, $msg, 'sent');
+			$tmp++;
+		}
+	}
+// Download messages if the download button was pressed
+	if (isset($_REQUEST["download"])) {
+		// if message ids are handed over, use them:
+		if (isset($_REQUEST["msg"])) {
+			foreach (array_keys($_REQUEST["msg"]) as $msg) {
+				$tmp = $messulib->get_message($user, $msg, 'sent');
+				$items[] = $tmp;
+			}
+		} else {
+			$items = $messulib->get_messages($user, 'sent', '', '', '');
+		}
+		$smarty->assign_by_ref('items', $items);
+		header("Content-Disposition: attachment; filename=tiki-msg-sent-" . time("U") . ".txt ");
+		$smarty->display('messu-download.tpl', null, null, null, 'application/download');
+		die;
+	}
 }
+
 if (isset($_REQUEST['filter'])) {
 	if ($_REQUEST['flags'] != '') {
 		$parts = explode('_', $_REQUEST['flags']);
@@ -108,6 +110,5 @@ $smarty->assign('cellsize', $cellsize);
 $smarty->assign('percentage', $percentage);
 include_once ('tiki-section_options.php');
 include_once ('tiki-mytiki_shared.php');
-ask_ticket('messu-sent');
 $smarty->assign('mid', 'messu-sent.tpl');
 $smarty->display("tiki.tpl");
